[ET-157] refactor(payment) : 패키지 구조 변경
